services/ingestion-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
aiokafka
asyncpg
redis
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
